Can now add and remove properly

+++ src/components/List.js
@@ -1,10 +1,19 @@
-import React, { Component, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import apiAddress from '../config/api';
 import ListItem from './ListItem';
+import Button from 'react-bootstrap/Button';
+import {BiMessageSquareAdd} from "react-icons/bi";
 
 const state = {
     fetched: false,
     items: []
+}
+
+const newItem = {
+    name: "",
+    details: "",
+    completed: 0,
+    image: null
 }
 
 const List = () => {
@@ -41,32 +50,48 @@ const List = () => {
     }
 
     const deleteItem = async (id) => {
-        await fetch(apiAddress + "/deleteItemById/" + String(id), {
-            method: "DELETE"
-        });
+        // Check if item is new item not registered in the database.
+        if (id) {
+            await fetch(apiAddress + "/deleteItemById/" + String(id), {
+                method: "DELETE"
+            });
+        }
 
-        const newItems = list.items.filter(listItem => listItem.id != id);
+        const newItems = items.filter(listItem => listItem.id !== id);
         setList({
             ...list,
             items: newItems
         });
     }
+
+    const createItem = () => {
+        items.push(newItem);
+
+        setList({
+            ...list,
+            items: items
+        });
+    }
     
-    if (!list.fetched) {
-        return (
-            <div>Loading...</div>
-        );
+    if (!fetched) {
+        return <div>Loading...</div>
     }
 
-    return (
-        <div className='List d-flex justify-content-center flex-wrap'>
+        return items.length !== 0 ?
+        (
+            <div className='List d-flex justify-content-center flex-wrap'>
             {
-                list.items.map((item) => (
+                items.map((item) => (
                     <ListItem item={item} checkboxHandler={changeCheckbox} deleteHandler={deleteItem} key={item.id}></ListItem>
                 ))
             }
-        </div>
-    );
+
+            <BiMessageSquareAdd size={30} onClick={createItem}></BiMessageSquareAdd>
+            </div>
+        )
+        :
+        
+        <Button variant="danger" onClick={createItem}>Make first item!</Button>
 }
  
 export default List;
